feat: remove the need to verify input
